Revert "New criteria for unanswered questions."
This reverts commit 0e839af287f1cd162ae477c60d30c31ad5148a7d.

The criterion of being "answered" when an answer has at least one
upvote was already implemented properly with the Question key
answered_with_id, and was broken with this commit (because it does the
update only on question *save*, not on new votes).

The problem we were seeing was because this after-vote-creation logic
is currently implemented in the controller, where it doesn't belong,
and the SE import script doesn't exercise any controller code.

So what we needed was to batch-update the answered_with status of
imported questions, and/or move this after-vote-creation logic to
models so that it would be executed during the import.