Change the way suggestions are generated on signup
When an external account is activated, suggestions are generated for
that external account (when applyed). When a user goes to the following
part of the signup, she'll wont have the topics from facebook, which is
a very time consuming task. It'll appear sometime, when the DJ that was
exectuted on the external account activation is done.